+++ aqua/util/__init__.py
@@ -4,7 +4,7 @@ from .config import ConfigPath
 from .eccodes import read_eccodes_dic, read_eccodes_def, get_eccodes_attr
 from .graphics import add_cyclic_lon, plot_box, minmax_maps
 from .graphics import evaluate_colorbar_limits, cbar_get_label, set_map_title
-from .graphics import coord_names, ticks_round
+from .graphics import coord_names, ticks_round, set_ticks
 from .sci_util import area_selection, check_coordinates
 from .util import generate_random_string, get_arg, create_folder, to_list
 from .util import file_is_complete, find_vert_coord
@@ -18,10 +18,11 @@ __all__ = ['ConfigPath',
            'read_eccodes_dic', 'read_eccodes_def', 'get_eccodes_attr',
            'add_cyclic_lon', 'plot_box', 'minmax_maps',
            'evaluate_colorbar_limits', 'cbar_get_label', 'set_map_title',
-           'coord_names', 'ticks_round',
-           'area_selection', 'check_coordinates', 'find_vert_coord', 'extract_literal_and_numeric',
-           'add_pdf_metadata', 'username',
-           'generate_random_string', 'get_arg', 'create_folder', 'to_list', 'file_is_complete',
+           'coord_names', 'ticks_round', 'set_ticks',
+           'area_selection', 'check_coordinates',
+           'generate_random_string', 'get_arg', 'create_folder', 'to_list',
+           'file_is_complete', 'find_vert_coord',
+           'extract_literal_and_numeric', 'add_pdf_metadata', 'username',
            'load_yaml', 'dump_yaml', 'load_multi_yaml', 'eval_formula',
            'check_chunk_completeness', 'frequency_string_to_pandas',
            'time_to_string',