+++ aqua/util/graphics.py
@@ -3,9 +3,13 @@ import math
 
 import xarray as xr
 import cartopy.util as cutil
+import cartopy.crs as ccrs
+import cartopy.mpl.ticker as cticker
 import numpy as np
+import matplotlib.pyplot as plt
 
 from aqua.logger import log_configure
+from .sci_util import check_coordinates
 
 
 def add_cyclic_lon(da: xr.DataArray):
@@ -82,8 +86,8 @@ def minmax_maps(maps: list):
         vmax (float): Maximum value of the colorbar.
     """
 
-    minmax = (min([map.min().values for map in maps]),
-              max([map.max().values for map in maps]))
+    minmax = (min([map.min().values for map in maps if map is not None]),
+              max([map.max().values for map in maps if map is not None]))
 
     vmin = minmax[0]
     vmax = minmax[1]
@@ -114,8 +118,8 @@ def evaluate_colorbar_limits(maps: list, sym: bool = True):
     if sym:
         vmin, vmax = minmax_maps(maps)
     else:
-        vmin = min([map.min().values for map in maps])
-        vmax = max([map.max().values for map in maps])
+        vmin = min([map.min().values for map in maps if map is not None])
+        vmax = max([map.max().values for map in maps if map is not None])
 
     return vmin, vmax
 
@@ -238,3 +242,78 @@ def ticks_round(ticks: list, round_to: int = None):
             round_to = 0
 
     return np.round(ticks, round_to)
+
+
+def set_ticks(data: xr.DataArray,
+              fig: plt.figure,
+              ax: plt.axes,
+              nticks: tuple,
+              lon_name: str,
+              lat_name: str,
+              ticks_rounding: int = None,
+              proj=ccrs.PlateCarree(),
+              loglevel='WARNING'):
+    """
+    Set the ticks of the map.
+
+    Args:
+        data (xr.DataArray): Data to plot.
+        fig (matplotlib.figure.Figure): Figure.
+        ax (matplotlib.axes._subplots.AxesSubplot): Axes.
+        nticks (tuple): Number of ticks for x and y axes.
+        lon_name (str): Name of the longitude coordinate.
+        lat_name (str): Name of the latitude coordinate.
+        ticks_rounding (int, optional): Number of digits to round the ticks.
+        loglevel (str, optional): Log level. Defaults to 'WARNING'.
+
+    Returns:
+        matplotlib.figure.Figure, matplotlib.axes._subplots.AxesSubplot: Figure and axes.
+    """
+    logger = log_configure(loglevel, 'set_ticks')
+    nxticks, nyticks = nticks
+
+    try:
+        lon_min = data[lon_name].values.min()
+        lon_max = data[lon_name].values.max()
+        (lon_min, lon_max), _ = check_coordinates(lon=(lon_min, lon_max),
+                                                  default={"lon_min": -180,
+                                                           "lon_max": 180,
+                                                           "lat_min": -90,
+                                                           "lat_max": 90},)
+        logger.debug("Setting longitude ticks from %s to %s", lon_min, lon_max)
+    except KeyError:
+        logger.critical("No longitude coordinate found, setting default values")
+        lon_min = -180
+        lon_max = 180
+    step = (lon_max - lon_min) / (nxticks - 1)
+    xticks = np.arange(lon_min, lon_max + 1, step)
+    xticks = ticks_round(ticks=xticks, round_to=ticks_rounding)
+    logger.debug("Setting longitude ticks to %s", xticks)
+    ax.set_xticks(xticks, crs=proj)
+    lon_formatter = cticker.LongitudeFormatter()
+    ax.xaxis.set_major_formatter(lon_formatter)
+
+    # Latitude labels
+    # Evaluate the latitude ticks
+    try:
+        lat_min = data[lat_name].values.min()
+        lat_max = data[lat_name].values.max()
+        _, (lat_min, lat_max) = check_coordinates(lat=(lat_min, lat_max),
+                                                  default={"lon_min": -180,
+                                                           "lon_max": 180,
+                                                           "lat_min": -90,
+                                                           "lat_max": 90},)
+        logger.debug("Setting latitude ticks from %s to %s", lat_min, lat_max)
+    except KeyError:
+        logger.critical("No latitude coordinate found, setting default values")
+        lat_min = -90
+        lat_max = 90
+    step = (lat_max - lat_min) / (nyticks - 1)
+    yticks = np.arange(lat_min, lat_max + 1, step)
+    yticks = ticks_round(ticks=yticks, round_to=ticks_rounding)
+    logger.debug("Setting latitude ticks to %s", yticks)
+    ax.set_yticks(yticks, crs=proj)
+    lat_formatter = cticker.LatitudeFormatter()
+    ax.yaxis.set_major_formatter(lat_formatter)
+
+    return fig, ax